feat(ui): Kept — bouncing wordmark + shared useSyncBounce hook [S]
The save-wave returns on the Kept brand: the boomerang. wordmark in the
mobile header and desktop sidebar renders per-letter spans driven by the
same sync states as the standard header (bounce on save, green flash on
sync, dim on degraded/offline), with the gold period riding the wave.

Dedupe: the hold/flash state machine was duplicated in Header.jsx and
WallabyHeader.jsx — extracted to src/hooks/useSyncBounce.js and consumed
by all three headers. syncStatus/queueLength threaded into KeptShell and
KeptDesktop.

Verified live: captured mid-wave in 'saving' on a real task completion;
idle -> saving -> idle cycle confirmed. Lint 0, build + tests pass.
